mymazon/requirements.txt
amqp==5.0.6
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-celery-results==2.2.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
dnspython==2.1.0
drf-spectacular==0.20.0
drf-spectacular-sidecar==2021.9.27
eventlet==0.32.0
greenlet==1.1.1
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.0.1
kombu==5.1.0
model-bakery==1.3.2
oauthlib==3.1.1
packaging==21.0
pluggy==1.0.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-django==4.4.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
toml==0.10.2
ujson==4.1.0
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
